link plugin executables statically
Previously the plugin executables installed by this script had dependencies
on shared libraries (`libpthread.so.0` and `libc.so.6`). These libraries
might not be installed on the kubelet host system, or might be at a different
version. If that's the case, the kubelet service won't be able to execute
these plugins.

Linking the executables statically is a more portable solution.